Centralize HandlePipelineProcessException between mmp and mtouch
